Formatting, remove comments, add shell script
Cleaned up code overall, unifying formatting and removing excess comments or commented out lines of code. Also added a shell script that binds the PHP development server to 0.0.0.0, allowing other computers on your LAN (or over the internet if you forward the port) to access your instance of the game.