First working version of randomiseObjects()
To remove the constant randomisation, comment out the contents of gameStep().
Ideally the values of x and z will be normalised to the coordinates of our boundary/plane. This can be done with the modulo operator and the addition of a constant, but I haven't done it yet.
The function iterates over an array in global scope called randomiserArray[]. Every model that's to be affected by the randomiser should be pushed to this array, by calling randomiserArray.push(modelname) at the end of its initialisation function (as is currently done with spawnChair()).
We can later implement player coordinate checks in the render loop and call this new function appropriately.