Added two different implementations of gameworld looping
The key functions for the tile-moving version are TerrainMatrix() and moveWithCamera(). For the player-moving version, they are checkBoundary() and updateWheels().
I got the tile-moving version working, but it isn't working 100% just yet and I also wanted to try my own idea, which is the player-moving version.
The player-moving version checks to see if the player is beyond the boundary of the plane's size (currently set to be 90% of the total size of the plane), and teleports them to the opposite side of the plane if so. This entails a whole lot of difficulties, as car.frame cannot be the parent of the wheels for some reason. This means that they are not moved when the trolley model is moved manually, and will consequently generate a massive amount of momentum if left unchecked.
The updateWheels() function is called by the main boundary checking function, and basically sums the Vector3 representing the current position of the trolley with another vector containing each wheel's offset from the trolley. This works, and somewhat solves the issue, but for some reason it causes the wheels to kind of 'roll' into place in an odd way (you'll see what I mean). This might actually be a desirable effect, depending.
We can explore each implementation further and see which is the most fruitful.