Moved new camera stuff to initScene(), added manual rotation
We're moving to a new camera style where the camera is a child of the trolley and sits in a fixed position relative to it. I've also added some basic orbiting along the y-axis, though it needs some kind of smoothing and ideally you would be able to rotate around the x/z axis - not sure which atm, but basically it would let you go around the trolley in a circle as well as up or down. We will also need to check for and prevent collisions between the camera, the world objects, and the infinite plane. The latter should be easy, the others less so.
I also added all the new objects to the randomiser array, and reenabled the contents of gameStep() to demonstrate this. They also now teleport anywhere within the plane no matter how high you scale it :)