Add text functionality, move trolley->camera parenting to initCamera()
These text boxes are automatically aligned to the centre bottom of the screen. The nuances of their formatting can be changed but atm they draw with black background, white monospaced text, centre aligned.
A global boolean, textDisplaying, controls text clearing. This boolean is set to true at the end of drawText(). In initTextListeners(), an event listener for the spacebar is set up to call the clearText() function if that boolean is set to true. The drawText() function itself will also call clearText() if that boolean is true.

You can test the new text function by calling drawText("with some message you want to display surrounded by quotes"). Press space to clear it.

The camera was a child of the goal Object3D, but as that object was needed only for lerping, we can just parent the trolley (car.frame) to the camera directly. I moved this to initCamera().