Beginnings of collision detection
Each Physijs box now has a .model field that identifies the visible object with which it's associated. For the moment handleCollision() will only print this model identifier to the console, but soon we could use a switch statement to trigger appropriate text boxes for each object.
We will probably also want a timer to make sure that the user doesn't get spammed with a million text boxes at once, so that'll need to be added soon.