Add some examples of text presentation modes
Iteration and die roll are represented here. Having game state-based triggers for certain dialogues would be the third major way of doing this I think. If we do that at any point, game state can be easily tracked with booleans, so we can really use whatever interactions we like as triggers.

Values we could track:
- playtime
- has player teleported
- has player interacted with each object? each type of object? one object but not another? etc…
- which quadrants of the gameworld the player has visited
definitely more, but im all out of energy for tonite :)