Beginnings of audio
You need to wait for a user gesture before you can play audio - thems the rules. So we'll need to implement a title screen or something we can use to get that first gesture. For now we can link it up to interaction overall, but I've actually left it commented out because it absolutely wrecks performance on my computer, more than this program already does.

Also I accidentally left the trolley physics box on and the trolley model off, so I switched those back around.