Add example on-theme text, add forced collision checking
If you're pushing one object continuously Physijs will only spring a collision event for the first collision. It won't report subsequent collisions unless you move away from the object first, which is clunky.

To get around this we hijack some of Physijs' internal values, namely the 'touches' array of the trolley object, which records the (Physijs internal) id/s of the object/s it's touching. We take this value at the time of the initial collision and compare that to its new value five seconds later. If the two values are the same, we know we can safely call handleCollision() on last_collided, since the Physijs id and the object are a pair/linked. It doesn't matter too much that we're just handling element zero of each array because handling more than one concurrently is impossible.