It actually is necessary to check each array element in touches
I thought it didn't matter but it does. In the old way, if an object happens to hit you at the same time as you're touching another object, forceRecheckCollision() will see that the first array element is still the same and call a handleCollision on it. If that object happens to bounce away you can end up with the wrong object producing another object's lines - not good. Checking all the array elements ensures that if an additional object enters the touches array no additional text will be shown until a new collision has occurred with just one object.